make the Tools class static
I know this is not good OOP practive. But who cares ? This project was 100% procedural, now the move to
OOP is slow but on the way. Static helper class might not be good oop
but it's not that bad I'd say. Anyway, there will be tons of refactoring
in the future, so it doesn't hurt.