#365 - Fix Fatal error when editing templates in user control panel
It ssems that the editing templates function beginning in line 75 has not yet been implemented using the Templates class. I temporarily fixed this, but I think this still has to be improved.

I imported PDO and called getConnection(). Additionally the redirect was not set, so the page did not redirect back to the user control panel.