fix bug where edit page is left open, session expires and data is lost
This bug happens if the user has no cookie, and leaves the edit page
open a very long time, the php session expires and the quicksave
function is not working anymore.

Now when that happens, the user is redirected to the login page, but the
edited content is saved in localStorage and can be recovered after
relogin.