Add blockchain timestamping with bloxberg.org (#2560)
Original PR (#2418) was made by @tnaegele and @patschwa. This is a complete rework, using the correct api and coding conventions, adding tests too.

It adds a new button in view mode "Add to blockchain", that will generate a pdf, make a hash, certify this hash into the blockchain, fetch a proof which is a pdf in a zip and add the timestamped pdf to that zip, and attach that zip to the entity.

Unlike "traditional" timestamping, an entity can be timestamped several times.

At the moment the request sometimes fail with an error: 422 Unprocessable entity, not a list. But this seems to be an intermittent error as redoing the request works.

There are no options to activate/deactivate that feature, but this can be added later if needed. At the moment I believe it's best to allow anyone to do it, as it doesn't require any account nor token to work.