Add possibility to push metadata via REST API
So far the JSON metadata of an item/experiment can be querried through
GET requests to the REST API. However, POSTing metadata into an existing
item/experiment is currently not possible.

This patch implements the logic needed to push JSON metadata into
items/experiments.

So far, any existing metadata will just be overwritten (just like with
title, date, and body parameters). It is thus left to the client to
retrieve existing metadata, merge it with the new one and push the
result to the API.

fixes #3069