use SITE_URL config value (#3323)
`SITE_URL` is now a mandatory env value that will endup in the `config.php` file, so the app knows what is the canonical URL without having to guess or do complicated gymnastics with ports and trusted proxies. Many apps have this setting (Discourse, Mattermost, ...) so it's not crazy to require it. We already have SERVER_NAME but this doesn't include the port, that could be custom, so having a complete URL is best.

`url` key from `Config` was not mandatory and could only be set after instance start, so it could sometimes not be reached. Also causes issues during site transfer, now it should be clear that this must be changed if a site is moved.

Works along https://github.com/elabftw/elabimg/commit/605e2b29c662336655534ea20b91ad02eb9a2b0f.

fix #3319