Remove the need for config.php file (#4064)
Previously, config.php was the file needed to run eLabFTW. It would define core settings like mysql host/password. In the container, this file was generated by the init script. Now the values are no longer in a php file defining global constants, but rather in the php-fpm config file so the values are available in env and can get fetched when needed by the code.

In this PR, bin/install start also migrates to bin/console db:install as it is no longer necessary to split this (was not before anyway).

A nice aspect is that we no longer need to fiddle with the config.php file for running tests.