remove deletable_xp and deletable_item
now with the soft delete it doesn't make sense to prevent users from
deleting things, it even has become a hassle for users. It is also the
kind of things that makes the user experience weird as in: how come I
cannot delete this? Why is this button disabled??

So it's best to simply remove it and simplify things.