Permissions related changes (#4820)
* remove deletable_xp and deletable_item

now with the soft delete it doesn't make sense to prevent users from
deleting things, it even has become a hassle for users. It is also the
kind of things that makes the user experience weird as in: how come I
cannot delete this? Why is this button disabled??

So it's best to simply remove it and simplify things.

* fix force_exp_tpl: behavior wasn't respected

and hide the no template buttons for items

* add the possibility to define permissions for the target entry

fix #4130

* fix dashboard create new not having category for resources

* remove a deletableXp if

* fix incorrect reload of items on switch layout.

fix #4809