maintenance: drop the old password and salt columns (#4935)
In 4.0.0 (June 16th 2021), modern password hashing algorithm was introduced.
password_hash() is used instead of salted shasum. A mechanism to
seamlessly update the old password to the new one was put in place.

Users who did not login since 4.0.0 will need to reset their password.
It is safe to assume most users will not be impacted by this change.