allow a valid user with no team to select a team
this could happen for instance after a successful SAML login, but then
the user quits the app before selecting a team, so the user is left
created but with no team, and this breaks things. So now the user can
come back later and finish setup their account because they'll be asked
for a team.