Import, export and other changes (#5123)
Here we go again, a huge PR. Anyway, the `hypernext` branch is a development branch, so better to merge things earlier rather than later. There are still some work to be done, fixes to make, but it's good enough to be merged in dev branch.

I'll try and summarize the changes anyway.

# Export changes

In the profile, a new page allows creating exports (eln, zip, pdf, json) in an asynchronous manner


This works by requesting the actual export to `bin/console export` through a unix socket, where a go program (in elabimg) is listening to the socket and will process requests one by one.

More importantly, now a .eln can includes entities of various types, experiments, resources or templates, and it will follow links to export everything, so the re-import is also dealing correctly with links. This support for linked entries during export/import was the last main major point during instance move.

There is also a way now to export a full team with `bin/console teams:export N`, and same with import. This will also create the users if needed, and re-import everything as before, more or less.

# Import changes

The "Import from file" functionality has moved next to the Export functionality in Profile:


Only .eln and .csv files can be imported. That means zip import has been dropped (see [reasons](https://github.com/elabftw/elabftw/issues/5121#issuecomment-2168210665)).

Again, there are things that still need to be ironed out. For instance, now the entries can contain a value for Category or Author, so the import interface should allow different behavior: overwrite or fill with given value if empty.

That means new API endpoints


# List attachments

In the profile, a new tab will list all the attachments from the user


# Unrelated changes

**BREAKING API CHANGE**: `/team_tags` has moved to `teams/{id}/tags`. One can replace any calls to `team_tags` with `teams/current/tags`. A helpful error message is displayed if the old endpoint is used.

* `sig_keys` endpoint moved into `users/{id}`
* in `bin/console`, factorize the sqlFs call
* add `profile.ts` for all the JS interactions on the profile page
* use yarn 4.2.2 instead of 4.0.0
* makebackupzip greatly simplified
* introduce EntitySlugs which are comprised of an entity type and an id and serve as targets for export actions
* add BaseQueryParams class for basic query params such as limit or order or sort
* add Invoker class that is able to talk to the socket
* add ActiveUser class that throws exception if they are archived
* allow copying files when duplicating an entry
* add UltraAdmin class that represents the command line user