add option to always display own entries
some users reported issues with the change introduced in 5.0 where
entries from other teams the user belonged were no longer shown. One
workaround was to add themselves in the "users" part of canread, and
then use scope::everything, but a better approach is to have a user
setting to toggle this behavior. It's off by default, but when switched
on, entries from current user will always be listed, regardless of the
team in which they are currently logged in.