Batch actions for Admins (#5148)
This PR adds a Batch tab in Admin panel to do actions based on a selection of:

* category / status of experiments
* category / status of resources
* users
* tags

It allows actions such as Lock, Unlock, Archive, Delete and setting the read or write permissions finely.

also add lock related columns to items_types, not because we need them, but to homogeneize with the rest. and fix timestamps.