Allow defining an URL to fetch XML IDP metadat (#5184)
This change allows a Sysadmin to define an URL corresponding to an XML description of one or several SAML2 IDP.

The **Refresh** button will get the file, parse it and populate the IDP list with entries from that file, while retaining information about its "source", so it can be refreshed later. You also have the possibility to **Enable all** or **Disable all** IDPs of that source/URL. It is also possible to toggle the `auto_refresh` attribute.

It is still possible to define an IDP manually, as before.

Because we can end up with thousands of IDPs, loading all of them like before would result in a 47 Mb html page! oO

So now a modal window is loaded on edit with the content of that IDP. There is also a filter input on top of the list to find an IDP more easily. And the "Enabled" attribute can be toggled directly from the list.

* add `idps_sources` API endpoint to deal with these URL + table `idps_sources`
* create a generic HTTP get/post class with guzzle, and use it for bloxberg, fixing bloxberg timestamping on the way
* add a `$requester` property to `Idps()` and deal with the permissions in the class instead of in the api class
* improve `Idps()` class by adding a private `create()` method and constants for defaults
* add two ligther readAll functions to avoid having to load too much data for the login page or sysconfig page
* add `bin/console idps:refresh` to loop over all sources with `auto_refresh = 1` => this will also be a daily cronjob in the container
* make `UltraAdmin` class `is_sysadmin = 1`
* add `Url2Xml` class to convert a URL to a DOMDocument with XML content
* add `Xml2Idps` class to parse an XML file and generate an array of IDPs ready to be created/updated
* use the same modal for create or edit IDP
* only display the SAML auth if `saml_toggle` is set but also there is at least one enabled IDP
* add search capabilities to the IDP select element, as we might have many now