fix: extra_fields: process values in toggle body (fix #5180)
In show mode, the toggle body would only show the ID of the linked
values in extra fields. This change will correctly display their title
or fullname if user.
Note: the links are not generated unlik view mode where twig is
processing the metadata. This will probably need to be homogeneized at
some point.