bug/medium: saml: user identity attributes update during saml
When using SAML auth, if the parameter "Allow users to change their
first name, last name, or email address" was inactive (not default),
then the user would encounter an error because the app would try and
update the user but have no right to do so.

This is fixed by making the update() method of Users public and using it
for trusted calls, and moving this check into the patch() method.