bug/major: saml: prevent empty Name attribute for requested attributes
In the SP Metadata XML (/metadata.php), a change has recently been made
to use the user defined attributes, instead of hardcoding them. But in
some case, the value was empty, not null, but an empty string, so the
Name attribute would be empty, and Shibboleth software that regularly
parses SP metadata would choke on it and everything would explode.

This change prevents having an empty Name attribute for the requested
attributes in SP metadata.