ux: idp: catch error with invalid input in idp edit modal
Error could be from an empty rollover cert, now at least an error is
displayed in the UI instead of simply being logged in console and no
feedback to user that the Save action didn't work.

Ideally, a better UX should be present to validate the form in a better
way, and visually indicate in which field the error is, and not close
the modal.