bug/trivial: calendar: creating a calendar item draws an empty event (#5551)
creating an event in the calendar caused an empty event waiting for the user to click elsewhere before refreshing the element (unselectAuto: true by default).
this fix allows the item to update instantly so that when you create your event, the title is directly set instead of an empty value.

fix #5265