bug/minor: entities: fix issue with user leaving team with created entities (#5651)
This change fixes an issue where for instance Resource Categories created by a user that left the team would disappear.

To trigger bug: create a Resource Category with user A in team A. Remove association between user A and team A. With another user in team A, verify that the created Resource Category isn't visible.

This fixes it by removing the users_id condition on the left join of users2teams table when reading entity.