users table and other things (#5796)
# User visible changes

## Table for users

It all started when I wanted to use ag-grid table for listing users. So let's start with this:

The third tab (Users) of Sysconfig panel and Admin panels is now listing users in a reactive table. This allows showing more columns, easier filtering on some attributes, instant search, and is just better than previous table.

The top search bar will search in every columns. Caveat: for searching "Is Sysadmin" or "2FA" or other binary column, one has to use 0 or 1. Couldn't figure out how to also include the actually shown text in the filtering.

To edit an user, double click the row and get a modal window with all you need.

## Archived status per team

A user can now be archived in a given team.

So archived state is no longer part of the users table, but associated in users2teams.

# Get the experiments categories, experiments status, resources categories (aka resources templates) and resources status out of the Admin panel
Now users can manage it themselves, unless Admin forbids it (not the default). They get their own page, too.

# New menu entries

## Misc visible changes

- Universign timestamping should be more robust: the date parsing will work with or without microseconds.
- add a `bookable` filter for resources
- add `$output` to sendEmails for notifications so it can better handle output in case of errors, as now it's not throwing an exception but logging error and progressing to next entry

# Dev stuff


## Groups

Get rid of table groups, because is_sysadmin is tied to user and is_admin is tied to user in team.

## Populate rework

The Populate class and command were reworked to make it easier to change: unit tests no longer fail when number of total users is changing or if name of teams change. We still have Toto, Titi, Tata and Tutu but for the rest it's random.

Add a `--fast` flag to the command so in CI we skip all the useless experiments generation. This brings down the populate step to 6 seconds! There is also a new `-i` flag to control number of entries to create for each user.

## Teams class

Add Teams in App and reshuffle things around a bit to execute less sql queries.

Remove need for bypassreadpermissions: have more public functions.

## Filenames changes

Change the function supposed to handle files when downloading them. fix #5783

## SCSS

- Start using newer syntax with namespaced variables basically.
- add order rule for csslint, also force usage of variables for colors

## Main text

Add more filters on css classes allowed, along with properties.


## Misc

- upgrade codeql action versions
- get rid of scrutinizer completely, we focus on circleci
- add devMode to App
- get rid of the `config.debug` column, only rely on `DEV_MODE` env var
- remove `includeArchived` from read-users (archived state is no longer tied to user)
- handle correctly an uninstalled instance where bin/console is called by chronos
- add BinaryValue Enum for things that are 0 or 1
- improve csv code by removing need to have getHeader
- remove old api_key header for bloxberg
- use mb_substr where it should
- log audit logs with info instead of notice
- count all entities regardless of their state in info endpoint
- remove populate-heavy.yml (use iter flag instead)
- add functions to get a random user or admin in a team
- modify how i18n is loaded: now import i18next from ./i18n


## TODO

- move the procedures into some sort of header file so they are always accessible in all schema files

# Fixes

fix #5779
fix #5783