bug/minor: inventory: fix inventory storage location listing missing
duplicate containers. If two or more containers have the same quantity
and unit, they would be collapsed.

Another approach was to use UNION ALL instead of UNION, but here
selecting a unique c2i.id and c2e.id was enough to deduplicate them. The
ALL was previously removed in another commit so we don't want to
re-introduce it.

Also add the id discriminator before the permissions checks in WHERE.