feat: prevent html tags being saved on firstname and lastname
this could theoritically be used to inject html in the email send to
Admins after local user register. Still hard to pull because of
character limit and the email is sent in text anyway, not html, and
email clients don't execute JS. But it doesn't hurt to strip all html
tags from this.