chore & fix: replace vulnerable xlsx package with @e965/xlsx & fix overwrite existing files
The original xlsx package has known vulnerabilities flagged by Snyk and other scanners.
Switching to @e965/xlsx resolves these security issues.
Also fix replacing an existing file with current sheet.
Behaviour to reproduce: By default, the "replace existing file" is disabled until you use "load in sheet editor" with a file that's in your uploads.
When you load a file from uploads, the "replace existing" uses the current upload's id to replace it. But if you import a new file or create a new sheet, that uploadId was still set, and then the "replace existing" would overwrite the existing data with some new sheet/import because of the id.
This is why I pass the setCurrentUploadId so I can reset it once we upload new files or create new sheets.