bug/medium: scheduler: non-admin users could no longer cancel a booking
fix #6073

handle it in typescript instead of sending a php array with the is_cancellable info, since we can already get that info on the modal.
The bug occured when we changed the 'item' param on the scheduler page, to handle items[] and so that "isBookable" array was outdated/empty.
This change uses the typescript to get that info event, and use it to toggle whether or not the cancel block should be shown