feat: exports: simplify team association in scheduler exports (#6175)
fix #6124
The scheduler report previously used UsersHelper to append team names for each user when generating CSV rows. This approach has been removed in favor of retrieving team names (for the event's origin and not the user's) directly through an SQL join on the teams table.
The new query in Scheduler now joins `teams` and selects `teams.name`, removing the need for UsersHelper and the additional processing step in MakeSchedulerReport.
This reduces redundant queries and simplifies report generation logic.