security: upgrade symfony/http-foundation to 6.4.29
Fix CVE-2025-64500
The Request class improperly interprets some PATH_INFO in a way that
leads to representing some URLs with a path that doesn't start with a /.
This can allow bypassing some access control rules that are built with
this /-prefix assumption.

Does it impacts eLabFTW? Unlikely. We're still going to patch it and
issue a release.