feat: tinymce: add "Save and Go Back" to toolbar
implements #6115
When editing long entries, users previously had to scroll to the bottom of the page to access the "Save and Go Back" button. This update adds a new "Save and Go Back" menu item to TinyMCE’s top toolbar under the "File" menu.
The new action triggers the existing save-and-redirect behavior, making it easier and faster to save and return to the entry view without scrolling. Improves editing efficiency for entries with large text bodies.