ux/minor: search: make placeholder text dynamic based on page title (#6225)
fix #6224
The main search bar placeholder was hardcoded to "Search in experiments", and "resources", which made it inaccurate on other pages (e.g., res/exp template views).
This update builds the placeholder dynamically using the translated page title, ensuring it correctly reflects the current section (e.g., "search in resources templates", "search in experiments templates").
However, it adds a need for translation for "search in" string.


---------

Co-authored-by: Nicolas CARPi <nico-git@deltablot.email>