bug/minor: ux/search: preserve active filters when performing a new search (#6294)
fix #6284
Previously, when selecting a filter from the dropdown (in navbar), performing a new search from the main search bar cleared contextual filters such as category, scope, or bookable, forcing users
to reselect them.

This update preserves those filters by appending their current values as hidden inputs in the main search form when a search is performed. The result is a smoother and more intuitive search experience that maintains the selected context.